requirements.txt
altgraph==0.17.3
astor==0.8.1
attrdict==2.0.1
Babel==2.11.0
bce-python-sdk==0.8.74
beautifulsoup4==4.11.1
blis==0.7.9
cachetools==5.2.0
catalogue==2.0.8
certifi==2022.12.7
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.6
confection==0.0.3
contourpy==1.0.6
cssselect==1.2.0
cssutils==2.6.0
cycler==0.11.0
cymem==2.0.7
Cython==0.29.32
decorator==5.1.1
dill==0.3.6
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.4.1/en_core_web_sm-3.4.1-py3-none-any.whl
et-xmlfile==1.1.0
fire==0.4.0
Flask==2.2.2
Flask-Babel==2.0.0
fonttools==4.38.0
future==0.18.2
google==3.0.0
idna==3.4
imageio==2.22.4
imgaug==0.4.0
itsdangerous==2.1.2
Jinja2==3.1.2
kiwisolver==1.4.4
langcodes==3.3.0
lmdb==1.4.0
lxml==4.9.1
MarkupSafe==2.1.1
matplotlib==3.6.2
multiprocess==0.70.14
murmurhash==1.0.9
networkx==2.8.8
numpy==1.23.5
opencv-contrib-python==4.6.0.66
opencv-python==4.6.0.66
openpyxl==3.0.10
opt-einsum==3.3.0
packaging==22.0
paddle-bfloat==0.1.7
paddle-python==0.6.0
paddleocr==2.6.1.1
paddlepaddle==2.4.1
pandas==1.5.2
pathy==0.10.1
pdf2docx==0.5.6
pefile==2022.5.30
Pillow==9.3.0
premailer==3.10.0
preshed==3.0.8
protobuf==3.20.0
pyclipper==1.3.0.post4
pycryptodome==3.16.0
pydantic==1.10.2
pyinstaller==5.7.0
pyinstaller-hooks-contrib==2022.14
PyMuPDF==1.20.2
pyparsing==3.0.9
PyQt5==5.15.7
PyQt5-Qt5==5.15.2
PyQt5-sip==12.11.0
python-dateutil==2.8.2
python-docx==0.8.11
pytz==2022.6
PyWavelets==1.4.1
pywin32-ctypes==0.2.0
rapidfuzz==2.13.6
requests==2.28.1
scikit-image==0.19.3
scipy==1.9.3
Shapely==1.8.5.post1
six==1.16.0
smart-open==6.2.0
soupsieve==2.3.2.post1
spacy==3.4.3
spacy-legacy==3.0.10
spacy-loggers==1.0.4
srsly==2.4.5
termcolor==2.1.1
thinc==8.1.5
tifffile==2022.10.10
tqdm==4.64.1
typer==0.7.0
typing_extensions==4.4.0
urllib3==1.26.13
visualdl==2.4.1
wasabi==0.10.1
Werkzeug==2.2.2
